requirements.txt
streamlit
pillow==9.5.0
reportlab